JWT na usuwanie plików

+++ app/app.py
@@ -54,17 +54,20 @@ def delete_file(id):
         return response, 403
 
     login = login_manager.getLogin(session_id)
-    token = create_delete_token(login)
 
     # Map id to file name
-    url = Config.API_URL + f"/files?user=jan"
+    token = create_list_token(login)
+    url = Config.API_URL + f"/files?user={login}&token={token}"
     r = requests.get(url)
+    if (r.status_code != 200):
+        flash("Nie udało się pobrać pliku", "alert-danger")
+        return redirect(url_for('index'))
+
     files = r.json()
     file_name = files[id]['fileName']
 
-    url = Config.API_URL + f"/files?user=jan&file={file_name}"
-    # url = urllib.parse.quote(url)
-    # print(url)
+    token = create_delete_token(login)
+    url = Config.API_URL + f"/files?user=jan&file={file_name}&token={token}"
     r = requests.delete(url)
 
     if r.status_code == 200:
